+++ dristhi-app/AndroidManifest.xml
@@ -157,7 +157,7 @@
                 android:screenOrientation="landscape"
                 android:theme="@android:style/Theme.Light.NoTitleBar.Fullscreen"/>
 
-        <activity android:name=".bidan.kartuibu.NativeKISmartRegisterActivity"
+        <activity android:name=".bidan.view.activity.NativeKISmartRegisterActivity"
                   android:screenOrientation="landscape"
                   android:theme="@android:style/Theme.Light.NoTitleBar.Fullscreen"/>
