+++ dristhi-app/src/main/java/org/ei/drishti/bidan/service/KartuIbuService.java
@@ -1,7 +1,7 @@
-package org.ei.drishti.bidan.kartuibu;
+package org.ei.drishti.bidan.service;
 
 import org.ei.drishti.AllConstants;
-import org.ei.drishti.domain.TimelineEvent;
+import org.ei.drishti.bidan.repository.AllKartuIbus;
 import org.ei.drishti.domain.form.FormSubmission;
 import org.ei.drishti.repository.AllTimelineEvents;
