+++ dristhi-app/src/main/java/org/ei/drishti/bidan/service/formSubmissionHandler/KartuIbuRegistrationHandler.java
@@ -1,20 +1,21 @@
-package org.ei.drishti.bidan.kartuibu;
+package org.ei.drishti.bidan.service.formSubmissionHandler;
 
+import org.ei.drishti.bidan.service.KartuIbuService;
 import org.ei.drishti.domain.form.FormSubmission;
 import org.ei.drishti.service.formSubmissionHandler.FormSubmissionHandler;
 
 /**
  * Created by Dimas Ciputra on 2/16/15.
  */
-public class KartuIbuCloseHandler implements FormSubmissionHandler {
+public class KartuIbuRegistrationHandler implements FormSubmissionHandler{
     private KartuIbuService kartuIbuService;
 
-    public KartuIbuCloseHandler(KartuIbuService kartuIbuService) {
+    public KartuIbuRegistrationHandler(KartuIbuService kartuIbuService) {
         this.kartuIbuService = kartuIbuService;
     }
 
     @Override
     public void handle(FormSubmission submission) {
-        kartuIbuService.closeKartuIbu(submission);
+        kartuIbuService.register(submission);
     }
 }
