+++ dristhi-app/src/main/java/org/ei/drishti/bidan/view/contract/KartuIbuClients.java
@@ -1,4 +1,4 @@
-package org.ei.drishti.bidan.kartuibu;
+package org.ei.drishti.bidan.view.contract;
 
 import org.ei.drishti.view.contract.SmartRegisterClients;
