+++ dristhi-app/src/main/java/org/ei/drishti/bidan/view/dialog/WifeAgeSort.java
@@ -1,4 +1,4 @@
-package org.ei.drishti.bidan.kartuibu;
+package org.ei.drishti.bidan.view.dialog;
 
 import org.ei.drishti.Context;
 import org.ei.drishti.R;
